add facility ID to location
