+++ data-query/src/test/java/gov/va/api/health/dataquery/service/controller/location/DatamartLocationTest.java
@@ -4,6 +4,7 @@ import static gov.va.api.health.autoconfig.configuration.JacksonConfig.createMap
 import static java.util.Optional.empty;
 import static org.assertj.core.api.Assertions.assertThat;
 
+import gov.va.api.health.dataquery.service.controller.FacilityId;
 import gov.va.api.lighthouse.datamart.DatamartReference;
 import java.util.Optional;
 import lombok.SneakyThrows;
@@ -31,13 +32,21 @@ public class DatamartLocationTest {
                 .reference(Optional.of("390026:I"))
                 .display(Optional.of("OLIN E. TEAGUE VET CENTER"))
                 .build())
+        .facilityId(
+            Optional.of(
+                FacilityId.builder()
+                    .stationNumber("623GB")
+                    .type(FacilityId.FacilityType.HEALTH)
+                    .build()))
         .build();
   }
 
   @SneakyThrows
-  private void assertReadable(String json) {
+  private static void assertReadable(String json) {
     DatamartLocation dm =
-        createMapper().readValue(getClass().getResourceAsStream(json), DatamartLocation.class);
+        createMapper()
+            .readValue(
+                DatamartLocationTest.class.getResourceAsStream(json), DatamartLocation.class);
     assertThat(dm).isEqualTo(sample());
   }
 
@@ -45,6 +54,7 @@ public class DatamartLocationTest {
   public void lazy() {
     DatamartLocation dm = DatamartLocation.builder().build();
     assertThat(dm.description()).isEqualTo(empty());
+    assertThat(dm.facilityId()).isEqualTo(empty());
     assertThat(dm.type()).isEqualTo(empty());
     assertThat(dm.physicalType()).isEqualTo(empty());
   }
